refactored service.go splitting into directors.go and handlers.go
