+++ go.mod
@@ -3,7 +3,7 @@ module github.com/ONSdigital/florence
 go 1.13
 
 require (
-	github.com/ONSdigital/dp-api-clients-go v1.23.0
+	github.com/ONSdigital/dp-api-clients-go v1.24.0
 	github.com/ONSdigital/dp-healthcheck v1.0.5
 	github.com/ONSdigital/dp-net v1.0.6
 	github.com/ONSdigital/dp-s3 v1.5.1
